pyproject.toml
[project]
name = "guardian"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "docx2txt>=0.9",
    "faiss-cpu>=1.13.2",
    "gradio>=6.3.0",
    "langchain-aws>=1.2.1",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.7",
    "langgraph>=1.0.6",
    "mcp>=1.25.0",
    "pushover-complete>=2.0.0",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.1",
]
